Added arm64 support. Changed deployment target to 5.1.1.
Fixes #178

+++ Demo/PhotoBrowserDemo/Menu.m
@@ -1114,7 +1114,7 @@
 //}
 
 - (void)photoBrowser:(MWPhotoBrowser *)photoBrowser didDisplayPhotoAtIndex:(NSUInteger)index {
-    NSLog(@"Did start viewing photo at index %i", index);
+    NSLog(@"Did start viewing photo at index %lu", (unsigned long)index);
 }
 
 - (BOOL)photoBrowser:(MWPhotoBrowser *)photoBrowser isPhotoSelectedAtIndex:(NSUInteger)index {
@@ -1123,7 +1123,7 @@
 
 - (void)photoBrowser:(MWPhotoBrowser *)photoBrowser photoAtIndex:(NSUInteger)index selectedChanged:(BOOL)selected {
     [_selections replaceObjectAtIndex:index withObject:[NSNumber numberWithBool:selected]];
-    NSLog(@"Photo at index %i selected %@", index, selected ? @"YES" : @"NO");
+    NSLog(@"Photo at index %lu selected %@", (unsigned long)index, selected ? @"YES" : @"NO");
 }
 
 #pragma mark - Load Assets
